+++ src/game/WorldSocket.cpp
@@ -343,7 +343,7 @@ int WorldSocket::handle_output (ACE_HANDLE)
 
         return -1;
     }
-    else if (n < send_len) //now n > 0
+    else if (n < (ssize_t)send_len) //now n > 0
     {
         m_OutBuffer->rd_ptr (static_cast<size_t> (n));
 
@@ -400,7 +400,7 @@ int WorldSocket::handle_output_queue (GuardType& g)
         mblk->release();
         return -1;
     }
-    else if (n < send_len) //now n > 0
+    else if (n < (ssize_t)send_len) //now n > 0
     {
         mblk->rd_ptr (static_cast<size_t> (n));
